Don't do any magic LO OS X application stuff in a LOKit client
(cherry picked from commit 20c9ceb708e4720205a303bf2d41274b1311e36b)

+++ vcl/osx/salinst.cxx
@@ -23,6 +23,8 @@
 
 #include <tools/solarmutex.hxx>
 
+#include "comphelper/lok.hxx"
+
 #include "osl/process.h"
 
 #include "rtl/ustrbuf.hxx"
@@ -182,6 +184,9 @@ static void initNSApp()
 
 bool ImplSVMainHook( int * pnInit )
 {
+    if (comphelper::LibreOfficeKit::isActive())
+        return false;
+
     unlink([[NSString stringWithFormat:@"%@/Library/Saved Application State/%s.savedState/restorecount.plist", NSHomeDirectory(), MACOSX_BUNDLE_IDENTIFIER] UTF8String]);
     unlink([[NSString stringWithFormat:@"%@/Library/Saved Application State/%s.savedState/restorecount.txt", NSHomeDirectory(), MACOSX_BUNDLE_IDENTIFIER] UTF8String]);
